refactor: renamed and merged files in "services" to api
